+++ Classes/Backend/DashboardPresetsProvider.php
@@ -15,7 +15,9 @@ use Brotkrueml\MatomoWidgets\Adapter\ExtensionAvailability;
 use Brotkrueml\MatomoWidgets\Configuration\Configuration;
 use Brotkrueml\MatomoWidgets\Configuration\ConfigurationFinder;
 use Brotkrueml\MatomoWidgets\Extension;
+use TYPO3\CMS\Core\Configuration\Loader\YamlFileLoader;
 use TYPO3\CMS\Core\Core\Environment;
+use TYPO3\CMS\Core\Utility\GeneralUtility;
 
 /**
  * @internal
@@ -44,7 +46,12 @@ final class DashboardPresetsProvider
      */
     public function __construct($configurationFinder = null)
     {
-        $this->configurationFinder = $configurationFinder ?? new ConfigurationFinder(Environment::getConfigPath(), new ExtensionAvailability());
+        $this->configurationFinder = $configurationFinder
+            ?? new ConfigurationFinder(
+                Environment::getConfigPath(),
+                new ExtensionAvailability(),
+                GeneralUtility::makeInstance(YamlFileLoader::class)
+            );
     }
 
     public function getPresets(): array